+++ source/isaaclab/isaaclab/envs/utils/camera_view.py
@@ -18,6 +18,7 @@ from pxr import Sdf
 
 import isaaclab.sim as sim_utils
 from isaaclab.sensors.camera import Camera, CameraCfg
+from isaaclab.sim.views import FrameView
 
 _GENERATED_CAMERA_NAME = "VisualizerCamera"
 VISUALIZER_TILED_CAMERA_MAX_TILES = 100
@@ -225,10 +226,6 @@ def prim_world_positions(
     Uses ``FrameView`` first so PhysX/Fabric-backed transforms are current; falls
     back to USD only if the backend view cannot be constructed.
     """
-    from pxr import UsdGeom
-
-    from isaaclab.sim.views import FrameView
-
     xform_cache = UsdGeom.XformCache()
     positions = []
     try:
