+++ source/isaaclab_experimental/isaaclab_experimental/envs/mdp/__init__.py
@@ -9,18 +9,6 @@ This package forwards all stable MDP terms from :mod:`isaaclab.envs.mdp`, but ov
 functions with Warp-first implementations from :mod:`isaaclab_experimental.envs.mdp.rewards`.
 """
 
-# Forward stable MDP terms (commands/observations/terminations/etc.) but *exclude* rewards and actions.
-# Rewards and actions are provided by this experimental package to keep Warp-first execution.
-from isaaclab.envs.mdp.commands import *  # noqa: F401, F403
-from isaaclab.envs.mdp.curriculums import *  # noqa: F401, F403
-from isaaclab.envs.mdp.events import *  # noqa: F401, F403
-from isaaclab.envs.mdp.observations import *  # noqa: F401, F403
-from isaaclab.envs.mdp.recorders import *  # noqa: F401, F403
-from isaaclab.envs.mdp.terminations import *  # noqa: F401, F403
+from isaaclab.utils.module import lazy_export
 
-# Override terms with experimental implementations.
-from .actions import *  # noqa: F401, F403
-from .events import *  # noqa: F401, F403
-from .observations import *  # noqa: F401, F403
-from .rewards import *  # noqa: F401, F403
-from .terminations import *  # noqa: F401, F403
+lazy_export()
