+++ source/isaaclab_experimental/isaaclab_experimental/envs/mdp/rewards.py
@@ -24,11 +24,10 @@ from isaaclab_newton.kernels.state_kernels import (
     rotate_vec_to_body_frame,
 )
 
-from isaaclab.assets import Articulation
-
 from isaaclab_experimental.managers import SceneEntityCfg
 
 if TYPE_CHECKING:
+    from isaaclab.assets import Articulation
     from isaaclab.envs import ManagerBasedRLEnv
