+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/direct/humanoid/humanoid_warp_env.py
@@ -5,106 +5,9 @@
 
 from __future__ import annotations
 
-from isaaclab_newton.physics import MJWarpSolverCfg, NewtonCfg
-
-import isaaclab.sim as sim_utils
-from isaaclab.assets import ArticulationCfg
-from isaaclab.envs import DirectRLEnvCfg
-from isaaclab.scene import InteractiveSceneCfg
-from isaaclab.sim import SimulationCfg
-from isaaclab.terrains import TerrainImporterCfg
-from isaaclab.utils.configclass import configclass
-
-from isaaclab_assets import HUMANOID_CFG
-
 from isaaclab_tasks_experimental.direct.locomotion.locomotion_env_warp import LocomotionWarpEnv
 
-
-@configclass
-class HumanoidWarpEnvCfg(DirectRLEnvCfg):
-    # env
-    episode_length_s = 15.0
-    decimation = 2
-    action_scale = 1.0
-    action_space = 21
-    observation_space = 75
-    state_space = 0
-
-    solver_cfg = MJWarpSolverCfg(
-        njmax=80,
-        nconmax=25,
-        cone="pyramidal",
-        integrator="implicitfast",
-        update_data_interval=2,
-        impratio=1,
-    )
-    newton_cfg = NewtonCfg(
-        solver_cfg=solver_cfg,
-        num_substeps=2,
-        debug_mode=False,
-        use_cuda_graph=True,
-    )
-
-    # simulation
-    sim: SimulationCfg = SimulationCfg(dt=1 / 120, render_interval=decimation, physics=newton_cfg)
-    terrain = TerrainImporterCfg(
-        prim_path="/World/ground",
-        terrain_type="plane",
-        collision_group=-1,
-        physics_material=sim_utils.RigidBodyMaterialCfg(
-            friction_combine_mode="average",
-            restitution_combine_mode="average",
-            static_friction=1.0,
-            dynamic_friction=1.0,
-            restitution=0.0,
-        ),
-        debug_vis=False,
-    )
-
-    # scene
-    scene: InteractiveSceneCfg = InteractiveSceneCfg(
-        num_envs=4096, env_spacing=4.0, replicate_physics=True, clone_in_fabric=True
-    )
-
-    # robot
-    robot: ArticulationCfg = HUMANOID_CFG.replace(prim_path="/World/envs/env_.*/Robot")
-    joint_gears: list = [
-        67.5000,  # left_upper_arm
-        67.5000,  # left_upper_arm
-        45.0000,  # left_lower_arm
-        67.5000,  # lower_waist
-        67.5000,  # lower_waist
-        67.5000,  # pelvis
-        45.0000,  # left_thigh: x
-        135.0000,  # left_thigh: y
-        45.0000,  # left_thigh: z
-        90.0000,  # left_knee
-        22.5,  # left_foot
-        22.5,  # left_foot
-        45.0000,  # right_thigh: x
-        135.0000,  # right_thigh: y
-        45.0000,  # right_thigh: z
-        90.0000,  # right_knee
-        22.5,  # right_foot
-        22.5,  # right_foot
-        67.5000,  # right_upper_arm
-        67.5000,  # right_upper_arm
-        45.0000,  # right_lower_arm
-    ]
-
-    heading_weight: float = 0.5
-    up_weight: float = 0.1
-
-    energy_cost_scale: float = 0.05
-    actions_cost_scale: float = 0.01
-    alive_reward_scale: float = 2.0
-    dof_vel_scale: float = 0.1
-
-    death_cost: float = -1.0
-    termination_height: float = 0.8
-
-    angular_velocity_scale: float = 0.25
-    contact_force_scale: float = 0.01
+from .humanoid_warp_env_cfg import HumanoidWarpEnvCfg
 
 
 class HumanoidWarpEnv(LocomotionWarpEnv):
