+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/locomotion/velocity/mdp/__init__.py
@@ -5,8 +5,6 @@
 
 """This sub-module contains the functions that are specific to the locomotion environments."""
 
-from isaaclab_experimental.envs.mdp import *  # noqa: F401, F403
+from isaaclab.utils.module import lazy_export
 
-from .curriculums import *  # noqa: F401, F403
-from .rewards import *  # noqa: F401, F403
-from .terminations import *  # noqa: F401, F403
+lazy_export()
