+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/locomotion/velocity/mdp/rewards.py
@@ -17,10 +17,10 @@ from typing import TYPE_CHECKING
 import warp as wp
 
 from isaaclab.managers import SceneEntityCfg
-from isaaclab.sensors import ContactSensor
 
 if TYPE_CHECKING:
     from isaaclab.envs import ManagerBasedRLEnv
+    from isaaclab.sensors import ContactSensor
 
 # ---------------------------------------------------------------------------
 # feet_air_time
