+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/manipulation/reach/mdp/__init__.py
@@ -5,6 +5,6 @@
 
 """This sub-module contains the functions that are specific to the reach environments."""
 
-from isaaclab_experimental.envs.mdp import *  # noqa: F401, F403
+from isaaclab.utils.module import lazy_export
 
-from .rewards import *  # noqa: F401, F403
+lazy_export()
